fix for isue #6: https://github.com/iamdustan/yellowbox-react/issues/6
Directly attaching an event handler that sets state will trigger thhis error. In my case using reactstrap’s <Modal> (https://reactstrap.github.io/components/modals/)
Example: <Modal toggle={this.toggleModal} /> where this.toggleModal() is *doing a this.setState()

What happens in my case is issue #6 in Yellowbox hides the source of the problem, in this case directly calling the Modal.toggle function when the component is mounting which causes a React.setStateLoopOfDoom error.

This patch makes sure that this._listener.remove() is a function before trying to call it, thereby allowing the real problem (state loop) to show up.

*Not sure it matters but I’m using immer to set state: https://github.com/immerjs/immer